Scenes&Rooms
In the environment game object, there are several rooms that can be used to design each level. To connect the rooms, you can delete tiles in the forground. All of the tiles have box colliders, so they can be placed anywhere in each room and the player can jump onto them. They can be found in the prefab folders along with the coins. The coins are destroyed upon collection.